kubernetes_agent/requirements.txt
uvicorn==0.35.0
langgraph==0.6.3
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.3
langgraph-sdk==0.2.0
langgraph-supervisor==0.0.29
langchain-core==0.3.72
langchain-mcp-adapters==0.1.9
langchain-ollama==0.3.6
langchain-openai==0.3.28
python-dotenv==1.1.1
a2a-sdk==0.3.0